Fix minor Apply iframe height issue.
In some situations, the Apply iframe is too short and it gets a nested scrollbar. It's most obvious when changing the "zoom" size or if viewing on mobile.

Fix changes:
- size to use `rem` vs `px` for browser zoom levels
- uses "mobile first" breakpoint that's taller for mobile and shorter for desktop.

No good way to test in localhost and staging because the iframe won't load. Attempted to test by applying new sizes to the production site via the debugger.

Rebuilding the css is problematic. Still must delete the style.css file and run the *release build* to get the correct output (not the `serve` or `start).