Downgrade to previous Ubuntu version, which is LTS
Recently, ubuntu-latest was updated to 24.04. This doesn't work with the version of setup-ruby that we use. Rather than debug that, downgrade to the previous version of Ubuntu which is LTS and supported for 2 years according to this doc: https://github.com/actions/runner-images/issues/10636.

setup-ruby should probably be updated at some point soon because our pinned version is 2+ years old.